requirements.txt
numpy == 1.10.2
pygubu == 0.9.7.8
Pillow == 3.0.0